Add timing thread pool to ZeroQueueConsumerImpl
