+++ pulsar-client-api/src/main/java/org/apache/pulsar/client/api/Consumer.java
@@ -119,9 +119,6 @@ public interface Consumer<T> extends Closeable {
      */
     Message<T> receive(int timeout, TimeUnit unit) throws PulsarClientException;
 
-
-    CompletableFuture<Message<T>> receiveAsync(int timeout,TimeUnit unit);
-
     /**
      * Batch receiving messages.
      *
