+++ pulsar-common/src/main/java/org/apache/pulsar/common/util/FutureUtil.java
@@ -23,7 +23,6 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 import java.util.function.Supplier;
 /**
  * This class is aimed at simplifying work with {@code CompletableFuture}.
